requirements.txt
adbutils==0.8.1
allure-pytest==2.7.0
allure-python-commons==2.7.0
apipkg==1.5
apkutils2==1.0.0
Appium-Python-Client==0.48
APScheduler==3.6.3
asgiref==3.3.1
assertpy==1.0
atomicwrites==1.4.0
attrdict==2.0.1
attrs==19.3.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.1
blinker==1.4
Brotli==1.0.7
bs4==0.0.1
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cigam==0.0.3
click==7.1.2
colorama==0.4.3
comtypes==1.1.7
cryptography==2.9.2
cycler==0.10.0
ddt==1.4.1
decorator==4.4.2
Deprecated==1.2.10
deprecation==2.1.0
Django==3.1.4
et-xmlfile==1.0.1
execnet==1.7.1
facebook-wda==1.0.5
Flask==1.1.2
gevent==20.9.0
greenlet==0.4.17
grequests==0.6.0
h11==0.10.0
h2==3.2.0
homebrew==0.1.13
hpack==3.0.0
humanize==2.5.0
hyperframe==5.2.0
idna==2.6
imageio==2.9.0
importlib-metadata==1.7.0
iniconfig==1.0.0
ipython==7.17.0
ipython-genutils==0.2.0
itchat==1.3.10
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
kaitaistruct==0.8
kiwisolver==1.2.0
ldap3==2.7
libimobiledevice==0.2
logzero==1.5.0
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mitmproxy==5.2
more-itertools==8.4.0
mysql==0.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.21
mysqlclient==2.0.1
networkx==2.4
npm==0.1.1
numpy==1.19.1
opencv-python==4.2.0.32
openpyxl==3.0.4
optional-django==0.3.0
packaging==20.4
pandas==1.1.0
paramiko==2.4.0
parso==0.7.1
passlib==1.7.2
pickleshare==0.7.5
pluggy==0.13.1
progress==1.5
prompt-toolkit==3.0.6
protobuf==3.11.3
publicsuffix2==2.20191221
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
pydivert==2.1.0
pyelftools==0.26
PyExecJS==1.5.1
Pygments==2.6.1
pymongo==3.10.1
PyMySQL==0.10.0
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.0
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.1.1.7.3
pytest==5.0.0
pytest-forked==1.3.0
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-ordering==0.6
pytest-rerunfailures==8.0
pytest-sugar==0.9.4
pytest-xdist==1.30.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyWavelets==1.1.1
pywin32==228
pywinauto==0.6.8
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
retry==0.9.2
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-image==0.16.2
scipy==1.5.2
selenium==3.141.0
six==1.15.0
sortedcontainers==2.1.0
soupsieve==2.0.1
sqlparse==0.4.1
termcolor==1.1.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
tzlocal==2.1
uiautomator2==2.10.1
Untitled==0.1
urllib3==1.22
urwid==2.0.1
wcwidth==0.2.5
weditor==0.5.0
Werkzeug==1.0.1
whichcraft==0.6.1
wrapt==1.12.1
wsproto==0.15.0
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
xmltodict==0.12.0
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.2.0
zstandard==0.13.0
